+++ rich/console.py
@@ -796,20 +796,18 @@ class Console:
         self,
         title: str = "",
         *,
-        character: Optional[str] = None,
         characters: str = "─",
         style: Union[str, Style] = "rule.line",
     ) -> None:
         """Draw a line with optional centered title.
         
         Args:
-            title (str, optional): Text to render over the rule. Defaults to "".
-            character: Will be deprecated in v6.0.0, please use characters argument instead.
+            title (str, optional): Text to render over the rule. Defaults to "".            
             characters (str, optional): Character(s) to form the line. Defaults to "─".
         """
         from .rule import Rule
 
-        rule = Rule(title=title, characters=character or characters, style=style)
+        rule = Rule(title=title, characters=characters, style=style)
         self.print(rule)
 
     def control(self, control_codes: Union["Control", str]) -> None: