+++ rich/text.py
@@ -644,7 +644,7 @@ class Text(JupyterMixin):
             length = cell_len(self.plain)
             if length > max_width:
                 if _overflow == "ellipsis":
-                    self.plain = set_cell_size(self.plain, max_width - 1).rstrip() + "…"
+                    self.plain = set_cell_size(self.plain, max_width - 1) + "…"
                 else:
                     self.plain = set_cell_size(self.plain, max_width)
             if pad and length < max_width: