+++ tests/test_console.py
@@ -294,3 +294,14 @@ def test_soft_wrap():
     console = Console(width=10, file=io.StringIO())
     console.print("foo bar baz egg", soft_wrap=True)
     assert console.file.getvalue() == "foo bar baz egg\n"
+
+
+def test_unicode_error() -> None:
+    try:
+        with tempfile.TemporaryFile("wt", encoding="ascii") as tmpfile:
+            console = Console(file=tmpfile)
+            console.print(":vampire:")
+    except UnicodeEncodeError as error:
+        assert "PYTHONIOENCODING" in str(error)
+    else:
+        assert False, "didn't raise UnicodeEncodeError"